+++ casm/vis/__init__.py
@@ -1,5 +1,11 @@
 """CASM visualizations."""
 
+from ._casm_vis import (
+    start,
+    stop,
+)
 from ._functions import (
     get_config,
+    get_pid_file,
+    get_root,
 )